Revert "Fix build on 64bit systems"
This reverts commit 1c6cef7bd2778ccecf9fec61afdbc293bde8bced.

When compiling on 64-bit systems, some of the fields of the serialized FDT
properties are not the proper width. The parsing code in GenericBooter uses
uint32_t type variables to represent the nProperties and nChildren fields in
the OpaqueDTEntry struct (see core/device_tree.c). I didn't dig deeply but
it seems fair to assume some part of dtc's codebase is using types with
arch-dependent widths.